store: tiny update
